+++ src/service/promql/engine.rs
@@ -1026,14 +1026,15 @@ async fn selector_load_data_from_datafusion(
             }
             MatchOp::Re(_re) => {
                 let regexp_match_udf =
-                    crate::service::search::datafusion::regexp_udf::REGEX_MATCH_UDF.clone();
+                    crate::service::search::datafusion::udf::regexp_udf::REGEX_MATCH_UDF.clone();
                 df_group = df_group.filter(
                     regexp_match_udf.call(vec![col(mat.name.clone()), lit(mat.value.clone())]),
                 )?
             }
             MatchOp::NotRe(_re) => {
                 let regexp_not_match_udf =
-                    crate::service::search::datafusion::regexp_udf::REGEX_NOT_MATCH_UDF.clone();
+                    crate::service::search::datafusion::udf::regexp_udf::REGEX_NOT_MATCH_UDF
+                        .clone();
                 df_group = df_group.filter(
                     regexp_not_match_udf.call(vec![col(mat.name.clone()), lit(mat.value.clone())]),
                 )?
