+++ src/service/search/datafusion/exec.rs
@@ -60,7 +60,7 @@ use once_cell::sync::Lazy;
 use parquet::arrow::ArrowWriter;
 use regex::Regex;
 
-use super::{storage::file_list, transform_udf::get_all_transform};
+use super::{storage::file_list, udf::transform_udf::get_all_transform};
 use crate::{
     common::meta::functions::VRLResultResolver,
     service::search::{datafusion::rewrite, sql::Sql, RE_SELECT_WILDCARD},
@@ -1209,23 +1209,23 @@ pub async fn prepare_datafusion_context(
 }
 
 async fn register_udf(ctx: &mut SessionContext, _org_id: &str) {
-    ctx.register_udf(super::match_udf::MATCH_UDF.clone());
-    ctx.register_udf(super::match_udf::MATCH_IGNORE_CASE_UDF.clone());
-    ctx.register_udf(super::regexp_udf::REGEX_MATCH_UDF.clone());
-    ctx.register_udf(super::regexp_udf::REGEX_NOT_MATCH_UDF.clone());
-    ctx.register_udf(super::regexp_udf::REGEXP_MATCH_TO_FIELDS_UDF.clone());
-    ctx.register_udf(super::time_range_udf::TIME_RANGE_UDF.clone());
-    ctx.register_udf(super::date_format_udf::DATE_FORMAT_UDF.clone());
-    ctx.register_udf(super::string_to_array_v2_udf::STRING_TO_ARRAY_V2_UDF.clone());
-    ctx.register_udf(super::arrzip_udf::ARR_ZIP_UDF.clone());
-    ctx.register_udf(super::arrindex_udf::ARR_INDEX_UDF.clone());
-    ctx.register_udf(super::arr_descending_udf::ARR_DESCENDING_UDF.clone());
-    ctx.register_udf(super::arrjoin_udf::ARR_JOIN_UDF.clone());
-    ctx.register_udf(super::arrcount_udf::ARR_COUNT_UDF.clone());
-    ctx.register_udf(super::arrsort_udf::ARR_SORT_UDF.clone());
-    ctx.register_udf(super::cast_to_arr_udf::CAST_TO_ARR_UDF.clone());
-    ctx.register_udf(super::spath_udf::SPATH_UDF.clone());
-    ctx.register_udf(super::to_arr_string_udf::TO_ARR_STRING.clone());
+    ctx.register_udf(super::udf::match_udf::MATCH_UDF.clone());
+    ctx.register_udf(super::udf::match_udf::MATCH_IGNORE_CASE_UDF.clone());
+    ctx.register_udf(super::udf::regexp_udf::REGEX_MATCH_UDF.clone());
+    ctx.register_udf(super::udf::regexp_udf::REGEX_NOT_MATCH_UDF.clone());
+    ctx.register_udf(super::udf::regexp_udf::REGEXP_MATCH_TO_FIELDS_UDF.clone());
+    ctx.register_udf(super::udf::time_range_udf::TIME_RANGE_UDF.clone());
+    ctx.register_udf(super::udf::date_format_udf::DATE_FORMAT_UDF.clone());
+    ctx.register_udf(super::udf::string_to_array_v2_udf::STRING_TO_ARRAY_V2_UDF.clone());
+    ctx.register_udf(super::udf::arrzip_udf::ARR_ZIP_UDF.clone());
+    ctx.register_udf(super::udf::arrindex_udf::ARR_INDEX_UDF.clone());
+    ctx.register_udf(super::udf::arr_descending_udf::ARR_DESCENDING_UDF.clone());
+    ctx.register_udf(super::udf::arrjoin_udf::ARR_JOIN_UDF.clone());
+    ctx.register_udf(super::udf::arrcount_udf::ARR_COUNT_UDF.clone());
+    ctx.register_udf(super::udf::arrsort_udf::ARR_SORT_UDF.clone());
+    ctx.register_udf(super::udf::cast_to_arr_udf::CAST_TO_ARR_UDF.clone());
+    ctx.register_udf(super::udf::spath_udf::SPATH_UDF.clone());
+    ctx.register_udf(super::udf::to_arr_string_udf::TO_ARR_STRING.clone());
 
     {
         let udf_list = get_all_transform(_org_id).await;
