+++ src/service/search/datafusion/udf/arrindex_udf.rs
