+++ src/service/search/datafusion/udf/arrzip_udf.rs
@@ -28,30 +28,30 @@ use datafusion::{
 use datafusion_expr::ColumnarValue;
 use once_cell::sync::Lazy;
 
-use crate::service::search::datafusion::stringify_json_value;
+/// The name of the arrzip UDF given to DataFusion.
+pub const ARR_ZIP_UDF_NAME: &str = "arrzip";
 
-/// The name of the arrjoin UDF given to DataFusion.
-pub const ARR_JOIN_UDF_NAME: &str = "arrjoin";
-
-/// Implementation of arrjoin
-pub(crate) static ARR_JOIN_UDF: Lazy<ScalarUDF> = Lazy::new(|| {
+/// Implementation of arrzip UDF
+pub(crate) static ARR_ZIP_UDF: Lazy<ScalarUDF> = Lazy::new(|| {
     create_udf(
-        ARR_JOIN_UDF_NAME,
-        // expects two string - the field and the delimiter
-        vec![DataType::Utf8, DataType::Utf8],
+        ARR_ZIP_UDF_NAME,
+        // expects three string - field1, field2 and the delim
+        vec![DataType::Utf8, DataType::Utf8, DataType::Utf8],
         // returns string
         Arc::new(DataType::Utf8),
         Volatility::Immutable,
-        Arc::new(arr_join_impl),
+        Arc::new(arr_zip_impl),
     )
 });
 
-/// arrjoin function for datafusion
-pub fn arr_join_impl(args: &[ColumnarValue]) -> datafusion::error::Result<ColumnarValue> {
-    log::debug!("Inside arrjoin");
-    if args.len() != 2 {
+/// arrzip function for datafusion
+pub fn arr_zip_impl(args: &[ColumnarValue]) -> datafusion::error::Result<ColumnarValue> {
+    log::debug!("Inside arrzip");
+    if args.len() != 3 {
         return Err(DataFusionError::SQL(
-            ParserError::ParserError("UDF params should be: arrjoin(field1, delim)".to_string()),
+            ParserError::ParserError(
+                "UDF params should be: arrzip(field1, field2, delim)".to_string(),
+            ),
             None,
         ));
     }
@@ -61,26 +61,35 @@ pub fn arr_join_impl(args: &[ColumnarValue]) -> datafusion::error::Result<Column
     // 1. cast both arguments to Union. These casts MUST be aligned with the signature or this
     //    function panics!
     let arr_field1 = as_string_array(&args[0]).expect("cast failed");
-    let delim = as_string_array(&args[1]).expect("cast failed");
+    let arr_field2 = as_string_array(&args[1]).expect("cast failed");
+    let delim = as_string_array(&args[2]).expect("cast failed");
 
     // 2. perform the computation
-    let array = zip(arr_field1.iter(), delim.iter())
+    let array = zip(arr_field1.iter(), zip(arr_field2.iter(), delim.iter()))
         .map(|(arr_field1, val)| {
             match (arr_field1, val) {
                 // in arrow, any value can be null.
                 // Here we decide to make our UDF to return null when either argument is null.
-                (Some(arr_field1), Some(delim)) => {
+                (Some(arr_field1), (Some(arr_field2), Some(delim))) => {
                     let arr_field1: json::Value =
                         json::from_str(arr_field1).expect("Failed to deserialize arrzip field1");
-                    let mut join_arrs = vec![];
-                    if let json::Value::Array(field1) = arr_field1 {
-                        field1.iter().for_each(|field| {
-                            let field = stringify_json_value(field);
-                            join_arrs.push(field);
+                    let arr_field2: json::Value =
+                        json::from_str(arr_field2).expect("Failed to deserialize arrzip field2");
+                    let mut zipped_arrs = vec![];
+                    if let (json::Value::Array(field1), json::Value::Array(field2)) =
+                        (arr_field1, arr_field2)
+                    {
+                        // Field1 and field2 can be of different types
+                        zip(field1.iter(), field2.iter()).for_each(|(field1, field2)| {
+                            let field1 = super::stringify_json_value(field1);
+                            let field2 = super::stringify_json_value(field2);
+
+                            zipped_arrs.push(format!("{field1}{delim}{field2}"));
                         });
                     }
-                    let join_arrs = join_arrs.join(delim);
-                    Some(join_arrs)
+                    let zipped_arrs =
+                        json::to_string(&zipped_arrs).expect("Failed to stringify zipped arrs");
+                    Some(zipped_arrs)
                 }
                 _ => None,
             }
@@ -107,58 +116,38 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn test_arr_join_udf() {
+    async fn test_arr_zip_udf() {
         let sqls = [
             (
                 // Should include values at index 0 to index 1 (inclusive)
-                "select arrjoin(bools, ',') as ret from t",
+                "select arrzip(bools, names, ',') as ret from t",
                 vec![
-                    "+-----------------+",
-                    "| ret             |",
-                    "+-----------------+",
-                    "| true,false,true |",
-                    "+-----------------+",
+                    "+-----------------------------------------+",
+                    "| ret                                     |",
+                    "+-----------------------------------------+",
+                    "| [\"true,hello2\",\"false,hi2\",\"true,bye2\"] |",
+                    "+-----------------------------------------+",
                 ],
             ),
             (
                 // Should include all the elements
-                "select arrjoin(names, ',') as ret from t",
+                "select arrzip(nums, floats, ',') as ret from t",
                 vec![
-                    "+-----------------+",
-                    "| ret             |",
-                    "+-----------------+",
-                    "| hello2,hi2,bye2 |",
-                    "+-----------------+",
+                    "+-----------------------------------------+",
+                    "| ret                                     |",
+                    "+-----------------------------------------+",
+                    "| [\"12,1.9\",\"345,34.5\",\"23,2.6\",\"45,4.5\"] |",
+                    "+-----------------------------------------+",
                 ],
             ),
             (
-                "select arrjoin(nums, ',') as ret from t",
+                "select arrzip(nums, mixed, ',') as ret from t",
                 vec![
-                    "+--------------+",
-                    "| ret          |",
-                    "+--------------+",
-                    "| 12,345,23,45 |",
-                    "+--------------+",
-                ],
-            ),
-            (
-                "select arrjoin(floats, ',') as ret from t",
-                vec![
-                    "+------------------+",
-                    "| ret              |",
-                    "+------------------+",
-                    "| 1.9,34.5,2.6,4.5 |",
-                    "+------------------+",
-                ],
-            ),
-            (
-                "select arrjoin(mixed, ',') as ret from t",
-                vec![
-                    "+---------------------------+",
-                    "| ret                       |",
-                    "+---------------------------+",
-                    "| hello2,hi2,123,23.9,false |",
-                    "+---------------------------+",
+                    "+--------------------------------------------+",
+                    "| ret                                        |",
+                    "+--------------------------------------------+",
+                    "| [\"12,hello2\",\"345,hi2\",\"23,123\",\"45,23.9\"] |",
+                    "+--------------------------------------------+",
                 ],
             ),
         ];
@@ -192,7 +181,7 @@ mod tests {
         // declare a new context. In spark API, this corresponds to a new spark
         // SQLsession
         let ctx = SessionContext::new();
-        ctx.register_udf(ARR_JOIN_UDF.clone());
+        ctx.register_udf(ARR_ZIP_UDF.clone());
 
         // declare a table in memory. In spark API, this corresponds to
         // createDataFrame(...).
